+++ tests/_bundles/myapp/config.py
@@ -1,7 +1,7 @@
-from flask_unchained import AppConfig
+from flask_unchained import AppBundleConfig
 
 
-class Config(AppConfig):
+class Config(AppBundleConfig):
     SECRET_KEY = 'change-me'
     APP_KEY = 'app_key'
     VENDOR_KEY1 = 'app_override'