+++ tests/_bundles/override_vendor_bundle/config.py
@@ -1,6 +1,6 @@
-from flask_unchained import AppConfig
+from flask_unchained import AppBundleConfig
 
 
-class Config(AppConfig):
+class Config(AppBundleConfig):
     VENDOR_KEY1 = 'override_vendor_key1'
     VENDOR_KEY2 = 'override_vendor_key2'